Project updated: - Removed otto and Butterknife - ViewBinding added
Tested in: Samsung S10+
Signed off by: Mauro Sarti mauro.sarti@globant.com
