+++ app/src/main/java/com/globant/mvpcounter/mvp/model/CountModel.java
@@ -1,4 +1,4 @@
-package com.globant.counter.android.mvp.model;
+package com.globant.mvpcounter.mvp.model;
 
 public class CountModel {
