+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
-    <string name="app_name">Android</string>
+    <string name="app_name">MVPCounter</string>
+    <string name="zero">0</string>
+    <string name="inc_button">inc</string>
+    <string name="reset_button">reset</string>
 </resources>
